+++ src/components/MobileMenu.tsx
@@ -3,7 +3,7 @@ import { usePathname, useSearchParams } from 'next/navigation';
 import { useEffect, useRef, useState } from 'react';
 import { RxHamburgerMenu } from 'react-icons/rx';
 
-import { Menu, TagWithCount } from '@/app/lib/types';
+import { Menu, TagWithCount } from '@/lib/types';
 import IconButton from '@/components/common/IconButton';
 import Sidebar from '@/components/Sidebar';
